feature: extend script to download deps. more embedability features.
For my workflows, it's really convenient to be able to take the entire
project directory and move it wholesale elsewhere (like the projects
section of my website.)

So now each template carries its own dependencies with it. Rather than
embedding those dependencies in this repository, when you download it
you can just run `./runner fetch` which will download all the dependencies
in deps.txt.

This is also convenient for offline development, provided you've
remembered to download those dependencies in advance.